comment out 3 layouts in gemm_dlops_int8; add 32x32x32 for gemm_dlops_int8; init A values to 1

+++ library/include/ck/library/utility/check_err.hpp
@@ -173,8 +173,8 @@ check_err(const std::vector<T>& out,
     {
         if(out[i] != ref[i])
         {
-            std::cout << "out[" << i << "] != ref[" << i << "]: " << out[i] << " != " << ref[i]
-                      << std::endl
+            std::cout << "out[" << i << "] != ref[" << i << "]: " << static_cast<int>(out[i])
+                      << " != " << static_cast<int>(ref[i]) << std::endl
                       << msg << std::endl;
             return false;
         }
